Split native settings view internals
