+++ Sources/QuillCodeApp/QuillCodeSettingsDraft.swift
@@ -1,0 +1,34 @@
+import Foundation
+import QuillCodeCore
+
+struct QuillCodeSettingsDraft: Equatable {
+    var apiBaseURL: String = ""
+    var authMode: TrustedRouterAuthMode = .oauth
+    var developerOverrideEnabled: Bool = false
+    var replacementAPIKey: String = ""
+    var shouldClearAPIKey: Bool = false
+
+    init() {}
+
+    init(settings: WorkspaceSettingsSurface) {
+        self.apiBaseURL = settings.apiBaseURL
+        self.authMode = settings.authMode
+        self.developerOverrideEnabled = settings.developerOverrideEnabled
+    }
+
+    var canSave: Bool {
+        !apiBaseURL.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
+    }
+
+    var update: WorkspaceSettingsUpdate {
+        WorkspaceSettingsUpdate(
+            apiBaseURL: apiBaseURL.trimmingCharacters(in: .whitespacesAndNewlines),
+            authMode: authMode,
+            developerOverrideEnabled: developerOverrideEnabled,
+            replacementAPIKey: replacementAPIKey.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
+                ? nil
+                : replacementAPIKey.trimmingCharacters(in: .whitespacesAndNewlines),
+            shouldClearAPIKey: shouldClearAPIKey
+        )
+    }
+}
