+++ Sources/QuillCodeApp/QuillCodeSettingsView.swift
@@ -127,256 +127,3 @@ struct QuillCodeSettingsView: View {
         }
     }
 }
-
-private struct QuillCodeComputerUseSettingsCard: View {
-    var settings: WorkspaceSettingsSurface
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            cardHeader
-            requirementRows
-            nextActionRow
-            restartHint
-            refreshAction
-        }
-        .padding(14)
-        .background(QuillCodePalette.panel.opacity(0.72))
-        .overlay(
-            RoundedRectangle(cornerRadius: 24, style: .continuous)
-                .stroke(statusTint.opacity(0.28), lineWidth: 1)
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 24, style: .continuous))
-        .shadow(color: Color.black.opacity(0.12), radius: 16, x: 0, y: 8)
-    }
-
-    private var statusTint: Color {
-        settings.computerUseStatus.available ? QuillCodePalette.green : QuillCodePalette.yellow
-    }
-
-    private var cardHeader: some View {
-        HStack(alignment: .top, spacing: 12) {
-            VStack(alignment: .leading, spacing: 4) {
-                Text("Computer Use")
-                    .font(.headline)
-                Text(settings.computerUseSetupSummary)
-                    .font(.caption)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            .layoutPriority(1)
-            Spacer()
-            Text(settings.computerUseStatusLabel)
-                .font(.caption.weight(.semibold))
-                .lineLimit(1)
-                .padding(.horizontal, 10)
-                .padding(.vertical, 6)
-                .background(statusTint.opacity(0.16))
-                .foregroundStyle(statusTint)
-                .clipShape(Capsule())
-        }
-    }
-
-    private var requirementRows: some View {
-        VStack(spacing: 8) {
-            ForEach(settings.computerUseRequirements) { requirement in
-                QuillCodePermissionRow(requirement: requirement, onCommand: onCommand)
-            }
-        }
-    }
-
-    private var nextActionRow: some View {
-        HStack(alignment: .top, spacing: 8) {
-            Image(systemName: settings.computerUseStatus.available ? "checkmark.circle.fill" : "arrow.forward.circle.fill")
-                .foregroundStyle(settings.computerUseStatus.available ? QuillCodePalette.green : QuillCodePalette.blue)
-                .frame(width: 18)
-            Text(settings.computerUseNextAction)
-                .font(.caption)
-                .foregroundStyle(settings.computerUseStatus.available ? QuillCodePalette.green : QuillCodePalette.muted)
-                .fixedSize(horizontal: false, vertical: true)
-                .layoutPriority(1)
-        }
-        .padding(10)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(QuillCodePalette.background.opacity(0.48))
-        .overlay(
-            RoundedRectangle(cornerRadius: 10, style: .continuous)
-                .stroke(Color.white.opacity(0.04), lineWidth: 1)
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-    }
-
-    private var restartHint: some View {
-        HStack(alignment: .top, spacing: 8) {
-            Image(systemName: "info.circle")
-                .foregroundStyle(QuillCodePalette.blue)
-                .frame(width: 18)
-            Text("After changing macOS permissions, quit and reopen QuillCode if the status does not update.")
-                .font(.caption2)
-                .foregroundStyle(QuillCodePalette.muted)
-                .fixedSize(horizontal: false, vertical: true)
-        }
-    }
-
-    private var refreshAction: some View {
-        HStack {
-            Button("Refresh status") {
-                onCommand(settings.computerUseRefreshCommand)
-            }
-            .buttonStyle(.borderless)
-            Spacer()
-        }
-        .font(.caption.weight(.semibold))
-    }
-}
-
-private struct QuillCodePermissionRow: View {
-    var requirement: ComputerUseRequirementSurface
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 12) {
-            ZStack {
-                Circle()
-                    .fill(iconTint.opacity(0.14))
-                Image(systemName: requirement.isGranted ? "checkmark.circle.fill" : "exclamationmark.circle.fill")
-                    .foregroundStyle(iconTint)
-            }
-            .frame(width: 28, height: 28)
-            VStack(alignment: .leading, spacing: 2) {
-                Text(requirement.title)
-                    .font(.callout.weight(.semibold))
-                Text(requirement.detail)
-                    .font(.caption)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            .layoutPriority(1)
-            Spacer(minLength: 12)
-            if requirement.isGranted {
-                Text(requirement.statusLabel)
-                    .font(.caption.weight(.semibold))
-                    .lineLimit(1)
-                    .foregroundStyle(QuillCodePalette.green)
-            } else {
-                Button("Open") {
-                    onCommand(requirement.command)
-                }
-                .buttonStyle(.bordered)
-                .disabled(!requirement.command.isEnabled)
-                .controlSize(.small)
-                .frame(minWidth: 72, minHeight: QuillCodeMetrics.minimumHitTarget)
-            }
-        }
-        .padding(10)
-        .background(Color.black.opacity(0.16))
-        .overlay(
-            RoundedRectangle(cornerRadius: 10, style: .continuous)
-                .stroke(Color.white.opacity(0.04), lineWidth: 1)
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-    }
-
-    private var iconTint: Color {
-        requirement.isGranted ? QuillCodePalette.green : QuillCodePalette.yellow
-    }
-}
-
-struct QuillCodeRuntimeIssueView: View {
-    var issue: RuntimeIssueSurface
-    var showsDiagnostics = false
-    var onAction: (() -> Void)?
-
-    var body: some View {
-        HStack(alignment: .top, spacing: 10) {
-            Image(systemName: issue.severity == .error ? "xmark.octagon.fill" : "exclamationmark.triangle.fill")
-                .foregroundStyle(tint)
-                .frame(width: 20)
-            VStack(alignment: .leading, spacing: 4) {
-                Text(issue.title)
-                    .font(.callout.weight(.semibold))
-                Text(issue.message)
-                    .font(.caption)
-                    .foregroundStyle(QuillCodePalette.muted)
-                if let actionLabel = issue.actionLabel {
-                    if let onAction {
-                        Button(actionLabel, action: onAction)
-                            .buttonStyle(.borderless)
-                            .font(.caption.weight(.semibold))
-                            .foregroundStyle(tint)
-                    } else {
-                        Text(actionLabel)
-                            .font(.caption.weight(.semibold))
-                            .foregroundStyle(tint)
-                    }
-                }
-                if showsDiagnostics && !issue.diagnostics.isEmpty {
-                    Divider()
-                        .padding(.vertical, 4)
-                    VStack(alignment: .leading, spacing: 5) {
-                        Text("Diagnostics")
-                            .font(.caption.weight(.semibold))
-                            .foregroundStyle(QuillCodePalette.muted)
-                        ForEach(issue.diagnostics) { diagnostic in
-                            HStack(alignment: .firstTextBaseline, spacing: 8) {
-                                Text(diagnostic.label)
-                                    .font(.caption2.weight(.semibold))
-                                    .foregroundStyle(QuillCodePalette.muted)
-                                    .frame(width: 96, alignment: .leading)
-                                Text(diagnostic.value)
-                                    .font(.caption2.monospaced())
-                                    .textSelection(.enabled)
-                                    .lineLimit(3)
-                            }
-                        }
-                    }
-                    .accessibilityElement(children: .contain)
-                }
-            }
-        }
-        .padding(12)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(tint.opacity(0.12))
-        .overlay(
-            RoundedRectangle(cornerRadius: 12, style: .continuous)
-                .stroke(tint.opacity(0.35), lineWidth: 1)
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
-    }
-
-    private var tint: Color {
-        issue.severity == .error ? QuillCodePalette.red : QuillCodePalette.yellow
-    }
-}
-
-struct QuillCodeSettingsDraft: Equatable {
-    var apiBaseURL: String = ""
-    var authMode: TrustedRouterAuthMode = .oauth
-    var developerOverrideEnabled: Bool = false
-    var replacementAPIKey: String = ""
-    var shouldClearAPIKey: Bool = false
-
-    init() {}
-
-    init(settings: WorkspaceSettingsSurface) {
-        self.apiBaseURL = settings.apiBaseURL
-        self.authMode = settings.authMode
-        self.developerOverrideEnabled = settings.developerOverrideEnabled
-    }
-
-    var canSave: Bool {
-        !apiBaseURL.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-
-    var update: WorkspaceSettingsUpdate {
-        WorkspaceSettingsUpdate(
-            apiBaseURL: apiBaseURL.trimmingCharacters(in: .whitespacesAndNewlines),
-            authMode: authMode,
-            developerOverrideEnabled: developerOverrideEnabled,
-            replacementAPIKey: replacementAPIKey.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-                ? nil
-                : replacementAPIKey.trimmingCharacters(in: .whitespacesAndNewlines),
-            shouldClearAPIKey: shouldClearAPIKey
-        )
-    }
-}
